LayoutTests/fast/tokenizer/ignore-tags-in-iframe.html

<html>
<head>
    <title></title>
    <script>
        if (window.testRunner)
            testRunner.dumpAsText();
    </script>
    <style>
        iframe { display: none; }
    </style>
</head>
<body>
    <p>
        Test for <i><a href="http://bugs.webkit.org/show_bug.cgi?id=12264">http://bugs.webkit.org/show_bug.cgi?id=12264</a>
        REGRESSION (r18843): Webkit crashes on 43Folders.com home page</i>.
    </p>
    <p>
        All tags below should report &ldquo;PASS&rdquo;
    </p>
    &lt;script>: <iframe><script></iframe>PASS</script></iframe>
    <br>
    &lt;style>: <iframe><style></iframe>PASS</style></iframe>
    <br>
    &lt;textarea>: <iframe><textarea></iframe>PASS</textarea></iframe>
    <br>
    &lt;title>: <iframe><title></iframe>PASS</title></iframe>
    <br>
    &lt;xmp>: <iframe><xmp></iframe>PASS</xmp></iframe>
</body>
</html>
