third_party/blink/web_tests/platform/generic/fast/tokenizer/ignore-tags-in-iframe-expected.txt

Test for http://bugs.webkit.org/show_bug.cgi?id=12264 REGRESSION (r18843): Webkit crashes on 43Folders.com home page.

All tags below should report “PASS”

<script>: PASS
<style>: PASS
<textarea>: PASS
<title>: PASS
<xmp>: PASS
